Extract memoizeOnSuccess helper
