+++ src/app/isNewSlug.ts
@@ -1,24 +1,7 @@
 import { client } from 'api/client';
 import query from './isNewSlugQuery.graphql';
 import { NotablePersonOldSlugQuery } from 'api/types';
-
-const memoizeOnSuccess = <T extends string, R>(
-  fn: ((arg: T) => Promise<R>),
-) => {
-  const map = new Map<string, R>();
-
-  return async (arg: T) => {
-    const cachedResult = map.get(arg);
-    if (cachedResult !== undefined) {
-      return cachedResult;
-    }
-
-    const result = await fn(arg);
-    map.set(arg, result);
-
-    return result;
-  };
-};
+import { memoizeOnSuccess } from 'helpers/memoizeOnSuccess';
 
 export const isNewSlug = memoizeOnSuccess(async (path: string) => {
   const result = await client.request<NotablePersonOldSlugQuery>(query, {
